refactor(frontend): colocate the home route and split the site shell
Move the home page under `(site)/(home)/` with its sections in `_components/`
and its copy and metadata in `_content.ts`, so a route owns the components only
it uses. Components shared by two or more routes stay in `src/components/`.

Turn `SiteShell` into a server component. Only the three parts that need
browser state remain client islands: `SiteNav` (active link and mobile menu,
kept together because they share the open state), `ThemeToggle` and
`SiteHeaderShell` (the entry animation alone, its children stay server-side).
The wordmark, the social links and the footer no longer ship as JavaScript.

Replace the hand-rolled theme script with `next-themes`. The theme icon is now
picked by CSS from `data-theme` instead of by JS after mount: the previous
mounted-gate rendered nothing on the server and an icon on the client, which
caused a hydration mismatch (React #418) on every page of the production build.

Also declare `eslint-import-resolver-typescript`, which is loaded by name
through the `import-x/resolver` setting and so escapes dependency analysis.
