+++ apps/frontend/package.json
@@ -32,6 +32,7 @@
     "lucide-react": "^0.460.0",
     "motion": "^13.1.0",
     "next": "^16.3.1",
+    "next-themes": "^0.4.6",
     "payload": "^3.88.0",
     "react": "^19.2.8",
     "react-dom": "^19.2.8",
