+++ apps/frontend/src/components/motion/primitives.tsx
@@ -5,6 +5,11 @@ import { m } from 'motion/react'
 import type { Variants } from 'motion/react'
 import type { ReactNode } from 'react'
 
+/*
+ * Motion takes easing as control points, CSS as a `cubic-bezier()` string, so the
+ * same curve has to exist in both forms. The CSS half lives in `globals.css` as
+ * `--ease-out-expo` / `--ease-out-quint`; changing one means changing the other.
+ */
 export const EASE_OUT_EXPO = [0.16, 1, 0.3, 1] as const
 export const EASE_OUT_QUINT = [0.22, 1, 0.36, 1] as const
