+++ apps/frontend/src/components/site-nav.tsx
@@ -1,0 +1,75 @@
+'use client'
+
+import { Menu, X } from 'lucide-react'
+import { m } from 'motion/react'
+import Link from 'next/link'
+import { usePathname } from 'next/navigation'
+import { useState } from 'react'
+
+const NAV_ITEMS = [
+  { href: '/', label: 'Accueil' },
+  { href: '/projets', label: 'Projets' },
+  { href: '/veille', label: 'Veille' },
+  { href: '/a-propos', label: 'À propos' },
+  { href: '/contact', label: 'Contact' },
+]
+
+/**
+ * Primary navigation and its mobile trigger.
+ *
+ * The links and the burger button are one component because they share the open
+ * state: below 900px the trigger is what shows the list. Splitting them would
+ * mean lifting that state into a context for no gain.
+ *
+ * The trigger renders after the links in the DOM but is placed in the header's
+ * action group by the parent, which passes it through as a separate slot.
+ */
+export function SiteNav({ actions }: { actions: React.ReactNode }) {
+  const pathname = usePathname()
+  const [menuOpen, setMenuOpen] = useState(false)
+
+  return (
+    <>
+      <nav
+        className={menuOpen ? 'nav-links is-open' : 'nav-links'}
+        aria-label="Navigation principale"
+      >
+        {NAV_ITEMS.map((item) => {
+          const active = pathname === item.href
+          return (
+            <Link
+              aria-current={active ? 'page' : undefined}
+              className={active ? 'nav-link is-active' : 'nav-link'}
+              href={item.href}
+              key={item.href}
+              onClick={() => setMenuOpen(false)}
+            >
+              {active ? (
+                <m.span
+                  className="nav-pill"
+                  layoutId="nav-active-pill"
+                  transition={{ type: 'spring', bounce: 0.16, duration: 0.55 }}
+                  aria-hidden="true"
+                />
+              ) : null}
+              <span className="active-dot" aria-hidden="true" />
+              <span className="nav-label">{item.label}</span>
+            </Link>
+          )
+        })}
+      </nav>
+      <div className="header-actions">
+        {actions}
+        <button
+          className="icon-link menu-trigger"
+          onClick={() => setMenuOpen((current) => !current)}
+          aria-label="Ouvrir le menu"
+          aria-expanded={menuOpen}
+          type="button"
+        >
+          {menuOpen ? <X size={19} /> : <Menu size={19} />}
+        </button>
+      </div>
+    </>
+  )
+}
