+++ packages/eslint-config/package.json
@@ -11,6 +11,7 @@
     "@next/eslint-plugin-next": "15.5.18",
     "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.0.0",
+    "eslint-import-resolver-typescript": "^3.0.0",
     "eslint-plugin-import-x": "^4.0.0",
     "eslint-plugin-react-hooks": "^5.0.0",
     "typescript-eslint": "^8.0.0"
